hunter2/privacy.py
# Copyright (C) 2021 The Hunter2 Contributors.
#
# This file is part of Hunter2.
#
# Hunter2 is free software: you can redistribute it and/or modify it under the terms of the GNU Affero General Public License as published by the Free
# Software Foundation, either version 3 of the License, or (at your option) any later version.
#
# Hunter2 is distributed in the hope that it will be useful, but WITHOUT ANY WARRANTY; without even the implied warranty of MERCHANTABILITY or FITNESS FOR A
# PARTICULAR PURPOSE.  See the GNU Affero General Public License for more details.
#
# You should have received a copy of the GNU Affero General Public License along with Hunter2.  If not, see <http://www.gnu.org/licenses/>.

from django.dispatch import receiver

from gdpr_assist.signals import pre_anonymise

from accounts.models import User


@receiver(pre_anonymise, sender=User)
def anonymise_allauth_models(sender, instance, **kwargs):
    instance.socialaccount_set.all().delete()
    instance.emailaddress_set.all().delete()
