pyproject.toml
[tool.poetry]
name = "hunter2"
version = "0.5.0"
description = "A platform for creating and running online or event-based puzzle hunts"
authors = ["The Hunter 2 contributors"]
license = "AGPL-3.0-or-later"

[tool.poetry.dependencies]
python = "^3.8"
Django = "==3.2.11"
PyYAML = "==6.0"
asgiref = "==3.4.1"  # Not really a direct dependency, included here to provide a version pin
channels = "2.4.0"
channels_redis = "==3.3.1"
daphne = "2.5.0"
django-allauth = "==0.47.0"
django-analytical = "==3.0.0"
django-autocomplete-light = "==3.8.1"
django-bootstrap4 = "==21.2"
django-debug-toolbar = "==3.2.2"
django-enumfields = "==2.1.1"
django-environ = "==0.8.1"
django-extensions = "==3.1.5"
django-fullclean = "0.0.5"
django-nested-admin = "==3.4.0"
django-ordered-model = { git = "https://github.com/fish-face/django-ordered-model.git", rev = "8a2908472b7bfbedb530a845f6c18fc86da7b750" }
django-prometheus = { git = "https://github.com/riconnon/django-prometheus.git", rev = "multi-process-dedicated-listener" }
django-redis-cache = "3.0.0"
django-seal = "==1.4.4"
django-solo = "==2.0.0"
django-sortedm2m = "==3.1.1"
django-tenants = "==3.4.1"
django-webpack-loader = "==1.4.1"
factory_boy = "==3.2.1"
freezegun = "1.0.0"
lupa = "==1.10"
parameterized = "==0.8.1"
psycopg2 = "==2.9.3"
python3-openid = "3.2.0"
rules = "==3.1"
sentry-sdk = "==1.5.2"
validators = "==0.18.2"
wait = "0.0.3"
unittest-xml-reporting = "3.0.4"
schema = "==0.7.5"
uWSGI = "==2.0.20"
django-sendfile2 = "0.6.0"
django-postgresql-dag = "==0.3.2"
pytest-django = "==4.5.2"
pytest-cov = "==3.0.0"
pytest-shard = "0.1.2"
django-gdpr-assist = { git = "https://github.com/wildfish/django-gdpr-assist.git", rev = "12a8f1e8dc0"}
dateparser = "^1.1.0"
[tool.poetry.dev-dependencies]
django-silk = "==4.2.0"
ipython = "==7.31.0"
ptpython = "==3.0.20"
pytest-profiling = "==1.7.0"
Sphinx = "==4.4.0"
myst-parser = "==0.16.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
